Made people activity and moved async task and recycler viewer to people activity, still need to tune up the layout
