+++ app/build.gradle
@@ -27,6 +27,7 @@ dependencies {
 
     implementation 'android.arch.persistence.room:runtime:1.1.1'
     implementation 'android.arch.lifecycle:extensions:1.1.1'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     annotationProcessor 'android.arch.persistence.room:compiler:1.1.1'
     annotationProcessor 'android.arch.lifecycle:compiler:1.1.1'
