+++ app/src/main/res/layout/activity_main.xml
@@ -57,9 +57,5 @@
         app:headerLayout="@layout/nav_drawer_header"
         app:menu="@menu/nav_drawer"/>
 
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/rv_starwars_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
 
 </android.support.v4.widget.DrawerLayout>
